add extra checks, remove current checks
