+++ style-guide/proselint-extra.py
@@ -1,13 +1,10 @@
-# -*- coding: utf-8 -*-
+"""Extra checks for proselint not generated by code parsing."""
 
-"""uk vs. us."""
-
-from proselint.tools import memoize, preferred_forms_check
-
+from proselint.tools import memoize, existence_check, preferred_forms_check
 
 @memoize
 def check_ukus(text):
-    """uk vs. us."""
+    """UK vs. US spelling usage."""
     err = "style-guide.uk-us"
     msg = "uk-vs-us-spell-check. '{}' is the preferred spelling."
 
@@ -1733,3 +1730,13 @@ def check_ukus(text):
     ]
 
     return preferred_forms_check(text, preferences, err, msg)
+
+@memoize
+def check_space(text):
+    """Use single space after sentence end."""
+    err = "style-guide.check-space"
+    msg = "Missing space after period."
+    regex = "[\.\?!][A-Z][a-z]"
+
+    return existence_check(text, [regex], err, msg, ignore_case=False, 
+                         require_padding=False)
