+++ style-test.txt
@@ -1,11 +1,2 @@
-        "style-guide.spelling-odk"      : true,
-        "style-guide.uk-us"             : true,
-        "style-guide.latin-abbr"        : true,
-        "style-guide.unneed-adverb"     : true,
-        "style-guide.filler-phrase"     : true,
-        "style-guide.personal-pronoun"  : true,
-        "style-guide.serial-comma"      : true,
-        "style-guide.check-same"        : true,
-        "style-guide.check-semicolon"   : true,
-        "style-guide.check-howto"       : true,
-        "style-guide.check-space"       : true
+        "style-guide.style-checks"      : true,
+        "style-guide.proselint-extra"   : true
